requirements.txt
# Code checkers
black==20.8b1
flake8==3.8.4
pre-commit==2.3.0

# Django
Django==3.1.3
djangorestframework==3.12.2
django-extensions==3.1.0
django-filter==2.4.0

# Postgres
psycopg2-binary==2.8.6

# Data pre-processing
# jupyter==1.0.0  # This is used for the data pre-processing in /notebooks.
# pandas==1.2.4  # This is used for the data pre-processing in  /notebooks.
